LayoutTests/http/tests/cookies/same-site/post-from-cross-site-popup-expected.txt


PASS '127.0.0.1' is same-site with itself, so Same-Site cookies are sent.
